+++ test/lib/BackRestTest/BackupTest.pm
@@ -1328,7 +1328,7 @@ sub BackRestTestBackup_Test
             #-----------------------------------------------------------------------------------------------------------------------
             $bDelta = true;
 
-            BackRestTestBackup_Restore($oFile, $strBackup, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+            BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                        undef, undef, undef, undef, undef, undef,
                                        'no tablespace remap', undef, '--no-tablespace', false);
 
@@ -1622,7 +1622,7 @@ sub BackRestTestBackup_Test
             $strComment = 'postmaster running';
             $iExpectedExitStatus = ERROR_POSTMASTER_RUNNING;
 
-            BackRestTestBackup_Restore($oFile, 'latest', $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+            BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                        $strType, $strTarget, $bTargetExclusive, $strTargetAction, $strTargetTimeline,
                                        $oRecoveryHashRef, $strComment, $iExpectedExitStatus);
 
@@ -1632,7 +1632,7 @@ sub BackRestTestBackup_Test
             $strComment = 'path not empty';
             $iExpectedExitStatus = ERROR_RESTORE_PATH_NOT_EMPTY;
 
-            BackRestTestBackup_Restore($oFile, 'latest', $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+            BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                        $strType, $strTarget, $bTargetExclusive, $strTargetAction, $strTargetTimeline,
                                        $oRecoveryHashRef, $strComment, $iExpectedExitStatus);
 
@@ -1646,7 +1646,7 @@ sub BackRestTestBackup_Test
             $strComment = undef;
             $iExpectedExitStatus = undef;
 
-            BackRestTestBackup_Restore($oFile, 'latest', $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+            BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                        $strType, $strTarget, $bTargetExclusive, $strTargetAction, $strTargetTimeline,
                                        $oRecoveryHashRef, $strComment, $iExpectedExitStatus);
 
@@ -1731,7 +1731,7 @@ sub BackRestTestBackup_Test
             $oFile->move(PATH_ABSOLUTE, BackRestTestCommon_TestPathGet() . '/recovery.conf',
                          PATH_ABSOLUTE, BackRestTestCommon_DbCommonPathGet() . '/recovery.conf');
 
-            BackRestTestBackup_Restore($oFile, 'latest', $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+            BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                        $strType, $strTarget, $bTargetExclusive, $strTargetAction, $strTargetTimeline,
                                        $oRecoveryHashRef, $strComment, $iExpectedExitStatus);
 
@@ -1808,7 +1808,7 @@ sub BackRestTestBackup_Test
 
                 BackRestTestBackup_ClusterStop();
 
-                BackRestTestBackup_Restore($oFile, 'latest', $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
+                BackRestTestBackup_Restore($oFile, OPTION_DEFAULT_RESTORE_SET, $strStanza, $bRemote, undef, undef, $bDelta, $bForce,
                                            $strType, $strTarget, $bTargetExclusive, $strTargetAction, $strTargetTimeline,
                                            $oRecoveryHashRef, $strComment, $iExpectedExitStatus);